update offer status
